Move subcomponents to their own folders
