+++ src/lib/sections/Navigation/Banner/Banner.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+
+export interface NavigationBannerProps {
+  children: ReactNode;
+}
+
+export const Banner = ({ children }: NavigationBannerProps) => {
+
+  return (
+    <>
+      {children}
+    </>
+  )
+}
