+++ src/lib/sections/Navigation/Header/Header.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+
+export interface NavigationHeaderProps {
+  children: ReactNode;
+}
+
+export const Header = ({ children }: NavigationHeaderProps) => {
+
+  return (
+    <div className="p-panel__header">
+      {children}
+    </div>
+  )
+}
