+++ src/lib/sections/Navigation/Item/Item.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from "react";
+
+export interface NavigationItemProps {
+  children: ReactNode;
+}
+
+export const Item = ({ children }: NavigationItemProps) => {
+  return (
+    <li className="p-side-navigation__item">
+      {children}
+    </li>
+  )
+}
