+++ src/lib/sections/Navigation/LogoName/LogoName.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+
+export interface NavigationLogoNameProps {
+  children: ReactNode;
+}
+
+export const LogoName = ({ children }: NavigationLogoNameProps) => {
+
+  return (
+    <div className="p-panel__logo-name is-fading-when-collapsed">
+      {children}
+    </div>
+  )
+}
