+++ src/lib/sections/Navigation/LogoTag/LogoTag.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+
+export interface NavigationLogoTagProps {
+  children: ReactNode;
+}
+
+export const LogoTag = ({ children }: NavigationLogoTagProps) => {
+
+  return (
+    <div className="p-navigation__logo-tag">
+      {children}
+    </div>
+  )
+}
